screwy woowy

+++ enc_temp_folder/d6d2fd5fc8df5956d9294355cbb3ada/Object.cpp
@@ -1,7 +1,7 @@
 #include "Object.h"
 #include <cstdlib>
 
-Object::Object(float x, float y, size_t ident) {
+Object::Object(float x, float y, float ident) {
 	pos.x = x;
 	pos.y = y;
 	color.r = 255;
