+++ enc_temp_folder/f68648dd1aa0736175c2bd825c4da5a3/Object.h
@@ -44,5 +44,5 @@ public:
 	bool isStatic;
 	Color color;
 
-	Object(float x, float y, size_t ident=0);
+	Object(float x, float y, float ident=0);
 };
